Reduced forward reasoning
A schematic rule which is a backward rule are not used as a forward
rule.

This cuts off many silly forward reasoning steps.

However some useful reasoning steps are cut off as well. E.g. the
assertion

    successor(n) = successor(m)

is not longer reduced to

    n = m

This deficiency will be removed in the future so that all simplifications
will be done properly.

	modified:   example/list_sort.al
	modified:   library/alba.base/boolean.al
	modified:   library/alba.base/boolean_logic.al
	modified:   library/alba.base/endorelation.al
	modified:   library/alba.base/function_logic.al
	modified:   library/alba.base/list.ali
	modified:   library/alba.base/natural.al
	modified:   library/alba.base/natural.ali
	modified:   library/alba.base/order_relation.ali
	modified:   library/alba.base/partial_order.al
	modified:   ocaml/proof_context.ml
	modified:   ocaml/proof_context.mli
	modified:   ocaml/proof_table.mli
	modified:   ocaml/rule_data.ml
	modified:   ocaml/source_prover.ml