Added evaluation term for beta reduction
Each beta reduction has a redex (reducible expression) and as a
result a reduct or beta contractum. For evaluation we need both
the redex and the reduct in unevaluated and evaluated form.

Constraint: The evaluated redex must be the unevaluated reduct.

Reason: In doing evaluation we can reach a point where we have to
evaluate a beta reducible expression. First we do the beta reduction
and then it might be possible to evaluate the reduct further.

	modified:   ocaml/proof.ml
	modified:   ocaml/proof_context.ml
	modified:   ocaml/proof_table.ml