Added parenthesis in syntax for nested induction proofs
In induction proofs there is a dangling 'case' problem

    inspect x
    case  ....
        assert
            ...
        inspect y
        case ...
        case ...
    case ...

There is the ambiguity if the last case belongs to the inner inspect or
to the outer inspect. By default the parsers groups the last case to
the inner inspect. If this is not the desired behaviour parentheses can
be added to enforce grouping.

    inspect x
    case ...
        (assert
             ...
         inspect y
         case ...
         case ... )
    case ...

	modified:   ocaml/parser.mly