Unfolding of inspect expressions no longer necessary
The symbolic evaluation of inspect expressions has been aligned
better with the evaluation at runtime.

The cases are analyzed one by one. For each case we have 3 possibilities.

a) The inspected expression can be matched with the pattern
b) The case can be rejected because of different constructors
c) The case cannot be decided

Actions:
a) Match: use this case
b) Reject: go to the next case
c) Undecidable: the whole inspect expression cannot be evaluated

In the case (c) it does not make sense to check the subsequent cases.
Even if we get a match it cannot be excluded that a previous case
can match as well.

	modified:   library/alba.base/natural.al
	modified:   library/alba.base/natural.ali
	modified:   ocaml/feature_table.ml
	modified:   ocaml/feature_table.mli
	new file:   ocaml/pattern.ml
	new file:   ocaml/pattern.mli
	modified:   ocaml/proof_context.ml
	modified:   ocaml/proof_table.ml
	modified:   ocaml/support.ml
	modified:   ocaml/term_algo.ml
	modified:   ocaml/term_algo.mli
	modified:   ocaml/typer.ml