Fix: Typing with ambiguous term applications wrong
In case that an expression can be used as a predicate or a function
the typer allowed wrong substitution of type variables.

In order to fix the bug the predicate/function ambiguity handling
has been redesigned.

If the application of a function/predicate term is analyzed the typer
checks if the result type can be of boolean type. If yes, the term
builders is duplicated and one tries to analyze the function term as
a function and the other as a predicate.

All handling of dummy types has been completely removed from the typer.

The error messages in case of ambiguities have been improved.

	modified:   ocaml/term_algo.ml
	modified:   ocaml/term_builder.ml
	modified:   ocaml/term_builder.mli
	modified:   ocaml/typer.ml