Fix: Don't evaluate the condition of an if expression
The condition of an if expression should not be evaluated during
term evaluation. Reason: It creates an asymmetry between the positive
and the negative case.

An if expression is evaluated if either the condition or its negation
is found in the context. If neither is found, the expression is not
evaluated because the branch cannot be decided.

	modified:   ocaml/proof_context.ml