Improved name clash resolution in printing terms
If there appears the same name in an inner and an outer context,
the name in the outer context is prefixed with '.'. If there are
already prefixed names then one more prefix is added.

I.e. the name '..v' does not refer to the name 'v' of the inner context,
it does refer to the name 'v' of the second outer context containing
a variable named 'v'.

	modified:   ocaml/feature_table.ml
	modified:   ocaml/feature_table.mli
	modified:   ocaml/proof_table.ml
	modified:   ocaml/term.ml