New finder for backward rules
The new finder for backward rules can now find also rules where the
target does not contain all variables. The missing variables are
found by searching for witnesses for the premises. However there are
not yet any backward rules in the tables which have targets with not
all variables. Therefore this part of the code is not yet tested.

Furthermore some hygiene changes are made:

1. Name change: Context.ntvs -> Context.count_all_type_variables

Reason: Clearer name

2. Table 'prvd' removed in proof context

Reason: The table is no longer in use.

	modified:   ocaml/context.ml
	modified:   ocaml/context.mli
	modified:   ocaml/feature_table.ml
	modified:   ocaml/proof_context.ml
	modified:   ocaml/proof_table.ml
	modified:   ocaml/rule_data.ml
	modified:   ocaml/term.ml
	modified:   ocaml/tvars.ml
	modified:   ocaml/tvars.mli