New finder for backward rules completed but not activated
The new finder for backward rules which can find backward rules with
targets which do not contain all variables has been completed.

However since there is no performace improvement but on the contrary
a performance degradation it has not been activated i.e. schematic
implications where the target does not contain all variables are not
treated as backward rules.

It seems that performance improvement by restricting forward rules
with partial specialization is more than compensated by the more
work it has to be done in backward reasoning.

	modified:   ocaml/rule_data.ml