Added declaration of a formal generic to the declaration of a deferred class
It is possible to declare a formal generic with the declaration of a
deferred class. Example:

    deferred class A:ANY end

This declaration is equivalent to the two declarations

    deferred class ANY end
    A: ANY

This change is convenient because any deferred classe needs a class
variable based on it. This is necessary because a deferred class cannot
be used directly but only via a type variable.

Furthermore this is a preparation for deferred inductive types like

    deferred class
        AL: ABSTRACT_LIST[A]
    create
        []
        (^) (hd:A, tl:AL[A])
    end

Without the simultaneous declaration of the class variable and the
deferred class it is not possible to refer to a class inheriting
ABSTRACT_LIST within the constructors.

	modified:   library/alba.base/any.al
	modified:   library/alba.base/any.ali
	modified:   ocaml/ast.ml
	modified:   ocaml/class_table.ml
	modified:   ocaml/class_table.mli
	modified:   ocaml/context.ml
	modified:   ocaml/context.mli
	modified:   ocaml/parser.mly
	modified:   ocaml/support.ml