Change the term representation of type terms
Types with actual generics are now represented as

    Application(Variable i, ags, false)

instead of

    VAppl(i,ags,[||],false)

This change is necessary in order to implement deferred classes with
formal generics. In this case we will have types of the form

    A[B,...]

where A is a formal generic. This formal generic has to be substitutable.
This is impossible with the representation VAppl(i,...) where 'i' cannot
be substituted.

	modified:   ocaml/ast.ml
	modified:   ocaml/class_table.ml
	modified:   ocaml/class_table.mli
	modified:   ocaml/context.ml
	modified:   ocaml/feature_table.ml
	modified:   ocaml/proof_context.ml
	modified:   ocaml/proof_table.ml
	modified:   ocaml/source_prover.ml
	modified:   ocaml/term.ml
	modified:   ocaml/term_builder.ml
	modified:   ocaml/tvars.ml
	modified:   ocaml/type_substitution.ml