Implemented concept of dominant class and dominant formal generic
A signature is owned by a class if it has exactly one involved class
which 'sees' all other involved classes. A class sees another class
if the other class has been declared before in the same module or
in a used module of the module of the class.

The owner class is also called the dominant class. If the signature
involves only classes declared in unrelated modules.

A dominant formal generic is the only formal generic of the signature
which is based on the dominant class of the signature.

All deferred features and deferred assertions must have a dominant
class and a dominant formal generic.

	modified:   ocaml/ast.ml
	modified:   ocaml/class_table.ml
	modified:   ocaml/class_table.mli
	modified:   ocaml/context.ml
	modified:   ocaml/context.mli
	modified:   ocaml/feature_table.ml
	modified:   ocaml/feature_table.mli
	modified:   ocaml/inherit.ml
	modified:   ocaml/module.ml
	modified:   ocaml/module.mli
	modified:   ocaml/proof_context.ml
	modified:   ocaml/proof_context.mli
	modified:   ocaml/prover.ml
	modified:   ocaml/source_prover.ml