Various fixes for inheritance and type unification
New rule:
    All variants of deferred features which substitute other formal
    generics than the dominant formal generic are illegal.

This rule is necessary and implemented.

Nearly all type unifications are done via Type_substitution. Only
the function Feature_table.unify_types is used for variant calculation
and the function Term_builder.unify is used for typing.

Both the variant calculation and the typing for deferred types with
formal generics i.e. with type variables which have formal generics
need to be revised because they are buggy in some situations.

	modified:   ocaml/ast.ml
	modified:   ocaml/class_table.ml
	modified:   ocaml/context.ml
	modified:   ocaml/context.mli
	modified:   ocaml/feature_table.ml
	modified:   ocaml/feature_table.mli
	modified:   ocaml/inherit.ml
	modified:   ocaml/type_substitution.ml
	modified:   ocaml/type_substitution.mli