New handling of class names
Different classes in different modules can have the same name.

Name conflicts with unqualified class names are resolved by giving
the class in the same module or in the same package precedence.

If a naming conflict cannot be resolved by the precedence rule then
the class name has to be qualified.

	modified:   ocaml/ast.ml
	modified:   ocaml/class_table.ml
	modified:   ocaml/class_table.mli
	modified:   ocaml/inherit.ml