Preliminary fix: Expression ambiguity
The typing of expressions can be ambiguous for 2 reasons.

1. It cannot be determined wheather a variable is of function or
   predicate type.

2. There are more possible functions calls at the same position in
   an expression and the typer cannot find out unambiguously which
   global function to call.

In the second case the different function calls are not printed
correctly if the different function calls are within a lambda
expression or an quantified expression and the calls involve bound
variables.

Reason: Term_algo.compare cannot find substitutions for the function
id because the function id in VAppl(id,...) cannot be substituted.

	modified:   ocaml/term_builder.ml