Improved generalization of induction goals
If the induction goal has variables beside the induction variable,
these are generalized. Example

    all(a,b,c:NATURAL)
        ensure
            a + b + c = a + (b + c)
        inspect
            c
        end

then the generalized induction goal

    all(a,b) a + b + c = a + (b + c)

is generated.

The induction hypotheses are entered in the generalized and the
specialized forms into the recursive induction cases.

Beside that this commit contains an improved finder for leibniz
terms which first searches for an equality term and then generates
the corresponding leibniz term.

	modified:   ocaml/context.ml
	modified:   ocaml/context.mli
	modified:   ocaml/proof_context.ml
	modified:   ocaml/proof_context.mli
	modified:   ocaml/source_prover.ml