Fix: Proof_context.add_to_proved
The function Rule_data.schematic_term does not work in the expected
way if the rule is already the target of implication chain and the
target is all quantified. Then rule date does not recognize the term
as a schematic term.

Therefore 'add_to_proved' cannot rely on Rule_data.

   (use "git push" to publish your local commits)