Added storage of induction laws in the class table
All induction laws (not only the induction laws of an inductive
class) are now recognized and stored within the corresponding class.

The induction law can be retrieved by giving a subset of the constructors.

This is support for pseudoinductive types. However induction laws for
pseudo inductive types are not yet used.

	modified:   ocaml/class_context.mli
	modified:   ocaml/class_table.ml
	modified:   ocaml/class_table.mli
	modified:   ocaml/container.ml
	modified:   ocaml/feature_context.ml
	modified:   ocaml/feature_context.mli
	modified:   ocaml/induction.ml
	modified:   ocaml/proof_context.ml
	modified:   ocaml/proof_context.mli
	modified:   ocaml/source_prover.ml
	modified:   ocaml/term.ml