Fix: No reordering of variables in automatic part of the prover
Reordering of variables and bubbling up of universal quantifiers only
happens for terms which occur in user proofs. The automatic part of
the prover escpecially the induction prover does not need and allow
any reordering and bubbling. Otherwise it may happen that some proved
premises have some unepected ordering of variables and universal
quantifiers.

	modified:   ocaml/feature_table.ml
	modified:   ocaml/proof_table.ml