Fix: Evaluation of inspect expressions
The evaluator iterates through all cases. A case is either undecidable,
accepted or rejected. If the case is undecidable, then the evaluation
stops and raise 'No_branch_evaluation'. If the case is accepted, then
the evaluation continues with the evaluation of this case. If a case
is rejected, then the next case is checked.

If all cases are rejected, then the cases are incomplete. This can happen
only if the expression has preconditions. But if the expression is
evaluated in a context with a contradiction then the precondition might
be violated and the evaluator falls through. In that case the evaluator
must raise 'No_branch_evaluation'.

	modified:   ocaml/proof_context.ml