Evaluation of inspected expressions is done lazily
Constructors don't have definitions and therefore cannot be evaluated.
If an inspected expression has a function with a definition at the
top it has to be evaluated recursively until an expression appears which
cannot be evaluated at the top level.

The evaluation of the inspected expression is done lazily without
evaluating its arguments.

If an if-expression is inside a case, then the if-expression sees the
arguments in unevaluted form. This is important especially for the
condition because the evaluator searches the positive or negative
condition in the context. The user usually states the condition in
unevaluated form.

	modified:   ocaml/proof_context.ml