Fix: Substitution of variable definitions
An assertion of the form 'v = exp' is a variable definition if 'v'
does not occur in 'exp', it has not yet a definition and none of the
variables in 'exp' have a definition.

A variable definition is used in all assertions in which it occurs to
substitute the variable by its definition term.

If the variables in 'exp' get a definition the definition has to be
applied to all assertions in which the definition 'v = exp' has already
been used.

	modified:   ocaml/proof_context.ml