Specialization of generalized induction hypotheses
If an induction proof has the form

    all(x,y,...)
        require
            r1; r2; ....
        ensure
            g
        inspect
            x
        ...
        end

then the goal predicate has the form

     p = {x: all(y,...) r1 ==> r2 ==> ... ==> g}

The induction hypotheses for a constructor have the form

    rai in p

where rai is a recursive argument of the constructor.

After beta reduction the induction hypothesis gets the form

    all(y,...) r1(rai) ==> r2(rai) ==> ... ==> g(rai)

If all the assumptions r1, r2, ... do not contain the induction variable
then the generalized induction hypothesis can be specialized by
using the variables y,... and the assumptions r1,r2,... of the inner
case context.

	modified:   ocaml/proof_context.ml
	modified:   ocaml/proof_context.mli
	modified:   ocaml/proof_table.mli
	modified:   ocaml/source_prover.ml