Fix: Prenex normal form for existential quantification
The variables in an existentially quantified expression shall be
sorted according to their appearance in the expression.

The number of bound variables have to be incremented appropriately.

Furthermore since existential quantification is used to code pattern
match expressions these existential expressions must not be normalized
since the appearance of the variables is significant (and there might
be missing variables which is not allowed in usual existential
quantification).

	modified:   ocaml/term.ml