Fix: Potential infinite loop in backward reasoning
Schematic rules which are not conservative (i.e. create subgoal which are more
complicated than the target) if reapplied have the potential to create endless
loops in the prover.

The current implementation prevents such potential endless loops by
maintaining a blacklist for each goal where the blacklist stores all non
conservative backward rules already used to generate it.

However this algorithm fails if a schematic rule is specialized in steps,
i.e. if the first step generates a schematic rule which is then directly or
indirectly used to generate more complex subgoals.

Now we store in each backward rule a complete list of schematic rules which
have been used to obtain it. If the rule is used to generate subgoals each of
the subgoals gets all used schematic rules on its blacklist. Therefore
reapplication of any of the schematic rules is prohibited.

	modified:   ocaml/proof_context.ml
	modified:   ocaml/proof_context.mli
	modified:   ocaml/prover.ml
	modified:   ocaml/rule_data.ml
	modified:   ocaml/rule_data.mli