Fix: Allow pattern match only if all requirements are satisfied
Pattern match is allowed on inductive types and on pseudo inductive
types.

For pseudo inductive types the following additional requirements must
be satisfied:

- All constructors have a complete set of projectors

- All constructors have recognizers which are mutually exclusive

The additional conditions for pseudo inductive types had been analyzed
but the result of the analysis had not been applied in the check of
pattern match expressions like inspect and as expressions.

	modified:   ocaml/class_table.ml
	modified:   ocaml/feature_table.ml