Pseudoinductive types use pattern matching with recognizers
An inspect expression with a pseudoinductive type is equivalent
to an if expression. E.g. for BOUNDED_NATURAL the function definition

    (<=) (a,b:N): BOOLEAN
        -> inspect a,b
           case 0, _ then
               true
           case _, 0 then
               false
           case n.successor, m.successor then
               n <= m

is equivalent to

    (<=) (a,b:N): BOOLEAN
        -> if a = 0 then
               true
           else if b = 0 then
               false
           else -- a /= 0 and b /= 0
               a.predecessor <= b.predecessor

The evaluator generates the conditions and tries to decide them.
In case that it cannot decide a precondition it fails with
No_branch_evaluation.

Optimazation hint: The guard condition for the last case needs no
proof since the other paths have already been decided and the cases
are always complete with respect to the preconditions.

	modified:   ocaml/context.ml
	modified:   ocaml/context.mli
	modified:   ocaml/feature_table.ml
	modified:   ocaml/feature_table.mli
	modified:   ocaml/induction.ml
	modified:   ocaml/pattern.ml
	modified:   ocaml/pattern.mli
	modified:   ocaml/proof_context.ml
	modified:   ocaml/proof_table.ml
	modified:   ocaml/term.ml
	modified:   ocaml/type_substitution.ml
	modified:   ocaml/type_substitution.mli