Fix: Reestablish the original enter sequence in the prover
If a goal is a universally quantified and/or an implication chain,
then the quantified variables and the premises of the chain are
pushed into the context.

However the target might be universally
quantified or an implication chain as well. Therefore the procedure
has to be iterated until a raw target appears which is neither
universally quantified nor an implication chain.

	modified:   ocaml/prover.ml