Removed some unnecessary checks for interface files
In interface files it is not necessary to check for termination of
recursive functions and to check that all used functions in pattern
are constructors.

This is done in the implementation file and the declarations in the
interface file have to be consistent with the declarations in the
implementation file.

	modified:   ocaml/ast.ml
	modified:   ocaml/term_builder.ml
	modified:   ocaml/typer.ml