Implementation of an inspect case hidden
An inspect case is implemented as a pair of lambda terms. One for
the constructor and one for the function which represents the case.

Both lambda terms must use the same variables. Therefore the actual
implementation has been hidden and a case is constructed by providing
the variable definitions and both inner expression.

Accessor function have been implemented to access the variable definitions,
the inner constructor term and the inner function term.

	modified:   ocaml/term2.ml
	modified:   ocaml/term2.mli
	modified:   ocaml/term_printer.ml
	modified:   ocaml/type_checker.ml