Refactored IO monad
There is now a module 'Io' which defines a minimal signature for an
IO module and a module which converts a module with a minimal
signature to a module with a complete signature.

The module 'Ocaml_io' only implements a minimal IO module and uses
Io.Make to convert the minimal module into a complete module.

	modified:   ocaml/alba2.ml
	modified:   ocaml/alba2_generic.ml
	modified:   ocaml/alba2_generic.mli
	new file:   ocaml/lib/io.ml
	new file:   ocaml/lib/io.mli
	modified:   ocaml/lib/ocaml_io.ml
	modified:   ocaml/lib/ocaml_io.mli