Simplified lookahead processing
Only 1 lookahead token needs to be processed on the second parser
of the >>= and <|> combinator if the second parser expects more
tokens.

The second parser does the same for each second step and therefore
either all lookahead tokens are processed or the parser terminates
and might leave tokens in the lookahead buffer.

	modified:   ocaml/lib/basic_parser.ml