Rework of 'Fmlib' and some implementation for node_js
Added content to Fmlib

- New monad interface  (return, >>=, <*>, >=>, ...)
- New module types FUNCTOR, APPLICATIVE, MONAD, all in Common_module_types
- List is now an own module in Fmlib
- Vector
- Pool
- Prepared streaming interface in 'Io' with 'READABLE', 'WRITABLE', 'FILTER',
  'String_reader', 'Fill_reader'
- Used a reference type to encapsulate the state in the 'State' monad.

New content in Fmlib.Node

The new library Fmlib.Node uses js_of_ocaml to wrap io functions of node_js
and make it available to ocaml.

This new content is an excercise on how to use js_of_ocaml i.e. access
javascript objects, wrap callbacks, call javascript functions etc.

- Basic filesystem access (open, read, write, rmdir, mkdir, ...)
- 'Process.exit(code)', 'Process.next_tick(cb)'
- io buffers
- module 'Node_io' which should satisfy the same interface as Ocaml_io. There
  is still work to be done.

Next steps

It turns out that the io functions like open, read, write, close are
relatively low level. Node has the concept of readable and writable streams
which is quite powerful.

If we have an incremental parser we could e.g. open a file as a readable
stream and parse it with the incremental parser in one command. The parser
just has to satisfy the interface of 'WRITABLE'. If we want as a side effect
to copy the file during parsing we can use the interface 'FILTER'. The
response of a http request can be fed into a parser etc.

Furthermore calling of external programs like `wget`, `curl`, `unzip`
sometimes require to read the output as a stream. Node has for each spawned
child process 3 streams: stdin, stdout and stderr.

Goal: Find a good high level interface for doing io, http requests and calling
external functions.