Added users to unfilled holes in 'Gamma_holes'
If an unfilled hole is filled, only its users must be expanded. This
makes filling of holes more efficient, because not all substitutions
must be resubstituted, only the substutitutions of the users.

The holes in the filling term get new users. The newly filled hole and
the users of the newly filled hole are new users for the unfilled holes
in the filling term.

	modified:   ocaml/albalib/gamma_holes.ml