Redesign of the unifier
The redesign has been necessary because of a bug.

The unifier added new local assumptions only to the 'Gamma' part and not
to the 'Gamma_holes' part. Therefore holes could not be expanded
properly ('Gamma' does not know the holes).

The unifier has now a monadic structure which leads to more readable
code.

	modified:   ocaml/albalib/builder.ml
	modified:   ocaml/core/gamma_algo.ml
	modified:   ocaml/core/gamma_algo.mli
	modified:   ocaml/core/gamma_holes.ml
	modified:   ocaml/core/gamma_holes.mli
	modified:   ocaml/core/unifier.ml
	modified:   ocaml/core/unifier.mli