Tried to integrate unification context into 'Gamma_holes'
But it did not work. 'Gamma_holes' gets confused if there are local
variable above the entries for the holes and the bound variables. The
index computation gets very complicated. Therefore it does not make
sense to integrate local variables above the holes and the bounds.

Better leave the extra needed local variables in a small wrapper in
'Unifier'.

	modified:   ocaml/albalib/build_context.ml
	modified:   ocaml/albalib/build_context.mli
	modified:   ocaml/albalib/builder.ml
	modified:   ocaml/core/gamma_holes.ml
	modified:   ocaml/core/gamma_holes.mli
	modified:   ocaml/core/typecheck.ml