Fmlib: Startedt improvement of pretty printer
The current implementation prints after each newline sufficient blanks
to fill the current indentation level. This is counterintuitive, if the
indentation ends immediately after the newline.

It is better to add the indentation blanks only at the output of some
indented text where the indentation level has not yet been reached. I.e.
check on each output if the current position is below the current
indentation level. If this is the case, then add the needed blanks.

The improvement has not yet been added, but a testcase has been
prepared.

	modified:   ocaml/fmlib/basic/pretty_printer.ml